requirements.txt
# API & HTTP
requests>=2.31.0
aiohttp>=3.9.0
xmltodict>=0.13.0

# Embeddings
sentence-transformers>=3.0.0
torch>=2.1.0
transformers>=4.36.0
peft>=0.10.0

# Vector Store
faiss-cpu>=1.7.4
numpy>=1.24.0

# Clustering
hdbscan>=0.8.33
scikit-learn>=1.3.0
umap-learn>=0.5.5

# RAG / LLM
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Data Processing
pandas>=2.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tqdm>=4.66.0

# Visualization / Dashboard
streamlit>=1.31.0
plotly>=5.18.0
matplotlib>=3.8.0
networkx>=3.2.0

# Utilities
tenacity>=8.2.0
